+++ pkg/adhoc/writer/external_writer.go
@@ -1,4 +1,4 @@
-package adhoc
+package writer
 
 import (
 	"fmt"
